refactor: split product-service into layers
