+++ product-service/lambdas/__tests__/getProductsList.test.js
@@ -1,6 +1,6 @@
 import { getProductsList } from '../getProductsList';
-import { successfulResponse, badResponse } from '../../helpers/responses';
-import { getPostgresClient } from '../../helpers/db';
+import { successfulResponse, badResponse } from '../../utils/responses';
+import { getPostgresClient } from '../../repositories/DynamoDB/helpers/getDynamodbClient';
 
 jest.mock('../../helpers/responses', () => ({
   successfulResponse: jest.fn(),
