+++ product-service/repositories/index.js
@@ -1,0 +1,2 @@
+export { DynamoDb } from './DynamoDB/DynamoDb';
+export { PostgreSQL } from './PostgreSQL/PostgreSQL';
