+++ product-service/services/ProductService.js
@@ -1,0 +1,11 @@
+export const ProductService = (repository) => {
+  const getAll = async () => repository.getAll();
+
+  const getById = async (id) => repository.getById(id);
+
+  const create = async (product) => repository.create(product);
+
+  const createBatch = async (records) => repository.createBatch(records);
+
+  return { getAll, getById, create, createBatch };
+};
