+++ product-service/utils/customError.js
@@ -1,6 +1,6 @@
 import { STATUS_CODE } from '../constants/statusCode';
 
-export const customError = (message, statusCode = STATUS_CODE.SERVER_ERROR) => {
+export const CustomError = (message, statusCode = STATUS_CODE.SERVER_ERROR) => {
   const error = new Error(message || 'Internal server error');
   error.statusCode = statusCode;
   return error;
